+++ main.py
@@ -397,7 +397,7 @@ def save_preferences(prefs):
 
 
 
-BROWSER_VERSION = "10.5.1"
+BROWSER_VERSION = "10.5.2"
 
 
 def _enable_per_monitor_dpi_awareness():